requirements.txt
flask
ddtrace
python-dotenv
